% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_case.R
\name{calibrate_case}
\alias{calibrate_case}
\title{Shiny gadget to select exam questions and generate both questions and solutions.}
\usage{
calibrate_case()
}
\value{
Save the different versions of exams and solutions.
}
\description{
Shiny gadget to select exam questions and generate both questions and solutions.
}
