% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_accounts.R
\docType{data}
\name{case_accounts}
\alias{case_accounts}
\title{List of accounts where transactions can be recorded for the case.}
\format{List of accounts where transactions can be recorded for the case.
\itemize{
\item case: id of the case
\item account: 5-digit account number. The first three digits indicate the nature, the last two the destination.
\item account_label: case-specific label for the account.
\item account_pool_object: cost pool or cost object or service to which the account is associated.
\item account_generic: generic label for the account.
\item account_subcategory: subcategory of the financial statements.
\item account_category: category of the financial statements.
\item account_statement: which financial statements.
}}
\usage{
data("case_accounts")
}
\description{
List of accounts where transactions can be recorded for the case.
}
\keyword{datasets}
