% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_activities.R
\docType{data}
\name{case_activities}
\alias{case_activities}
\title{List of activities classified per process.}
\format{List of activities classified per process.
\itemize{
\item case
\item activity_full: name of the family of products to which the prodict belongs.
\item activity_order: order in which the activity is conducted within the process.
\item activity_process: "design", "purchasing", "production", "selling", "administration".
\item activity_description: description of the input, steps and output of the activity.
}}
\usage{
data("case_activities")
}
\description{
List of activities classified per process.
}
\keyword{datasets}
