% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_companies.R
\docType{data}
\name{case_companies}
\alias{case_companies}
\title{List of general parameters for a simulation.}
\format{List of general parameters for a simulation.
\itemize{
\item case: id of the case.
\item company_type: indication about the profile of the company
\item number_cost_objects: number of different produts or services sold.
\item number_cost_pools: number of different cost pools used.
\item number_materials: number of different kinds of raw materials used.
\item number_joint_products: number of products which are joint.
}}
\usage{
data("case_companies")
}
\description{
List of general parameters for a simulation.
}
\keyword{datasets}
