% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_contracts.R
\docType{data}
\name{case_contracts}
\alias{case_contracts}
\title{List of contracts with the corresponding cost behavior and related taxes.}
\format{List of contracts with the corresponding cost behavior and related taxes.
\itemize{
\item case
\item contract_full: name of the contract.
\item contract_behavior: cost behavior resulting from the contract.
\item contract_driver: measure of activity to which the cost is proportional because of a contractual arrangement.
\item contract_description: description of the contract.
\item contract_premium: coefficient by which unit costs are multiplied to account for the risk implied by the contrat for the partner.
}}
\usage{
data("case_contracts")
}
\description{
List of contracts with the corresponding cost behavior and related taxes.
}
\keyword{datasets}
