% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_costing.R
\docType{data}
\name{case_costing}
\alias{case_costing}
\title{Cost assignment mapping.}
\format{Cost assignment mapping.
\itemize{
\item case: id of the case.
\item account: account from where the resources are drawn.
\item account_label: label of the account from where the resources are drawn.
\item allocation_base: measure of activity used to trace or allocate costs.
\item assignment: type of assignment: "tracing" or "allocation".
}}
\usage{
data("case_costing")
}
\description{
Cost assignment mapping.
}
\keyword{datasets}
