% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_families.R
\docType{data}
\name{case_families}
\alias{case_families}
\title{List of families of products}
\format{List of families of products
\itemize{
\item case
\item industry: Main industry to which the family belongs.
\item family_full: Full name of the family.
\item family_sing: Singularname of the family.
\item family_plur: Plural name of the family.
\item family_description: Description of the family of products.
\item base_volume: firm-day volume of reference for simulation.
\item sensitivity_prime_cost: customers' relative sensitivity to materials' value.
\item sensitivity_advertising: customers' relative sensitivity to advertising.
\item sensitivity_price: customers' relative sensitivity to price.
\item sensitivity_dso: customers' relative sensitivity to DSO (credit terms).
\item severance_cost: cost incurred for the termination of a permanent contract.
\item cancellation_cost: cost incurred for the termination of a lease.
\item asset_disposal_loss: percentage of the book value not recovered on asset disposal.
}}
\usage{
data("case_families")
}
\description{
List of families of products
}
\keyword{datasets}
