% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_industries.R
\docType{data}
\name{case_industries}
\alias{case_industries}
\title{List of industries and associated company names with initial financing policies.}
\format{List of industries and associated company names with initial financing policies.
\itemize{
\item case
\item industry
\item company: company name
\item country: country in which the firm has its headquarters.
\item age: Number o years since creation.
\item size: factor by which the family's base volume is multiplied.
\item desired_rate_of_return
\item interest_rate: initial interest rate.
\item overdraft_rate: interest rate applied on advances from the bank.
\item cost_of_capital: rate of return required by shareholders.
\item debt_share: proportion of the capital employed financed by debt at the beginning.
\item debt_years: number of years over which the debt was contracted.
\item value_added_tax: percentage of VAT on purchases.
\item labor_tax: percentage of tax applied on wages.
\item income_tax: percentage of EBT for the State.
\item dso: Days Sales Outstanding.
\item dio: Days Inventory Outstanding.
\item dpo: Days Payable Outstanding.
}}
\usage{
data("case_industries")
}
\description{
List of industries and associated company names with initial financing policies.
}
\keyword{datasets}
