% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_information.R
\docType{data}
\name{case_information}
\alias{case_information}
\title{List of case properties.}
\format{List of case properties.
\itemize{
\item case: id of the case.
\item number_cost_objects: number of different produts or services sold.
\item number_cost_pools: number of different cost pools used.
\item number_materials: number of different kinds of raw materials used.
\item number_joint_products: number of products which are joint.
\item type_costing: whether the company uses "actual", "normal" or "standard" costing.
\item allocation_method: whether the company uses the "direct", "sequential" or "reciprocal" method of allocation.
}}
\usage{
data("case_information")
}
\description{
List of case properties.
}
\keyword{datasets}
