% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_products.R
\docType{data}
\name{case_products}
\alias{case_products}
\title{List of products classified per family.}
\format{List of products classified per family.
\itemize{
\item case
\item family_full: Name of the family of products to which the prodict belongs.
\item product_full: Full name of the product.
\item product_sing: Singular name of the product.
\item product_plur: Plural name of the product.
\item product_variety: Number of different products in the family.
\item product_mono_material: Whether the product is made of only one type of raw material.
\item product_sequence: order in which products are made in a sequential process.
\item product_joint: whether the products within the same family oare joint.
\item product_price: initial price of the product on the market.
\item product_advertising: initial amount of monthly advertising.
}}
\usage{
data("case_products")
}
\description{
List of products classified per family.
}
\keyword{datasets}
