% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_resources.R
\docType{data}
\name{case_resources}
\alias{case_resources}
\title{List of presources consumed with their price}
\format{List of presources consumed with their price
\itemize{
\item case: id of the case.
\item resource: resource produced or consumed.
\item resource_unit: unit in which the resource is expressed.
\item resource_type: type of resource: "materials" ,"services", "utilities", "labor", "machines", "facilities" or "products".
}}
\usage{
data("case_resources")
}
\description{
List of presources consumed with their price
}
\keyword{datasets}
