% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_seasons.R
\docType{data}
\name{case_seasons}
\alias{case_seasons}
\title{Relative distribution of the demand over the year.}
\format{Relative distribution of the demand over the year.
\itemize{
\item case
\item week
\item monday
\item tuesday
\item wednesday
\item thursday
\item friday
\item saturday
\item sunday
}}
\usage{
data("case_seasons")
}
\description{
Relative distribution of the demand over the year.
}
\keyword{datasets}
