% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_technology.R
\docType{data}
\name{case_technology}
\alias{case_technology}
\title{Resources consumed at different phases.}
\format{Resources consumed by each product.
\itemize{
\item case: id of the case.
\item purpose: whether resources are consumed for "selling", "production", or "support".
\item input_standard_quantity: quantity of resource consumed to make this quantity of that output.
\item input_unit: unit of the resource consumed.
\item input: account number from which the resource comes.
\item input_label: label of the account from which the resource comes.
\item output_standard_quantity: quantity produced.
\item output_unit: unit of the quantity produced.
\item output: account number to which the resource goes.
\item output_label: label of the account to which the resource goes.
\item costing: whether the link between cost pools relates to tracing, accumulation or allocation.
\item from_pool: where the corresponding costs are first accumulated.
\item from_pool_label: where the corresponding costs are first accumulated.
\item to_pool: where the corresponding costs are then sent.
\item to_pool_label:  where the corresponding costs are then sent.
}}
\usage{
data("case_technology")
}
\description{
Resources consumed at different phases.
}
\keyword{datasets}
