% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_templates.R
\docType{data}
\name{case_templates}
\alias{case_templates}
\title{List of general parameters for a simulation.}
\format{List of general parameters for a simulation.
\itemize{
\item case: id of the case.
\item template: id of the template within the case.
\item mono_product: whether the case relies on a single product.
\item mono_material: whether the case relies on only one kind of raw materials.
\item joint_products: whether the case is about joint products.
\item costing: costing method used: "actual", "normal" or "standard".
\item allocation: allocation method used: "direct", "sequential" or "reciprocal".
\item inventory_fp_policy: whether finished products are managed to "smooth production" or in "just in time".
\item inventory_rm_policy: whether finished products are managed to "smooth purchases" or in "just in time".
\item desired_rate_of_return: minimum rate of return on assets desired.
\item interest_rate: initial interest rate.
\item overdraft_rate: interest rate applied on advances from the bank.
\item cost_of_capital: rate of return required by shareholders.
\item debt_share: proportion of the capital employed financed by debt at the beginning.
\item debt_years: number of years over which the debt was contracted.
\item value_added_tax: percentage of VAT on purchases.
\item labor_tax: percentage of tax applied on wages.
\item income_tax: percentage of EBT for the State.
\item base_volume: period volume serving as reference.
\item sensitivity_price: sales' sensitivity to price differences.
\item sensitivity_prime_cost: sales' sensitivity to prime costs (proxy for quality) differences.
\item sensitivity_dso: sales sensitivity to differences in credit terms.
\item sensitivity_advertising: sales sensitivity to differences in advertising expenses.
\item sensitivity_commissions: sales sensitivity to differences in sales commissions.
\item sensitivity_demand: how discriminant differences in price, cost, dso, advertising and commissions are.
\item downward_adjustment_cost: percentage of the yearly expense paid as penalty for decreasing capacity.
\item upward_adjustment_cost: percentage of the yearly expense paid as cost of increasing capacity.
}}
\usage{
data("case_templates")
}
\description{
List of general parameters for a simulation.
}
\keyword{datasets}
