% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coeffday.R
\docType{data}
\name{coeffday}
\alias{coeffday}
\title{Daily variations in market size}
\format{A data frame with 7 rows and 2 variables
\itemize{
\item weekday  Character. Days of the week.
\item coeffday Numeric. Magnitude of the market that day.
}}
\usage{
data("coeffday")
}
\description{
Daily variations in market size
}
\keyword{datasets}
