% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coeffmonth.R
\docType{data}
\name{coeffmonth}
\alias{coeffmonth}
\title{Monthly variations in market size}
\format{A data frame with 12 rows and 2 variables
\itemize{
\item month       Integer. Months
\item coeffmonth  Numeric. Magnitude of the market each month.
}}
\usage{
data("coeffmonth")
}
\description{
Monthly variations in market size
}
\keyword{datasets}
