% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_batches.R
\name{create_batches}
\alias{create_batches}
\title{Split a volume into dated batches.}
\usage{
create_batches(
  object = "lot",
  quantity = 1200,
  max_batch_nbr = 20,
  min_batch_size = 20,
  start_date = Sys.Date(),
  end_date = Sys.Date() + 20
)
}
\arguments{
\item{object}{Character. component of the final label identifying the batch.}

\item{quantity}{Integer. Number of units to be divided in batches.}

\item{max_batch_nbr}{Integer. Number of batches.}

\item{min_batch_size}{Integer. Minimum number of units in a batch.}

\item{start_date}{Date. Beginning of the period.}

\item{end_date}{Date. End of the period.}
}
\value{
A dataframe with date, object and quantity of each batch.
}
\description{
Split a volume into dated batches.
}
