% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_case.R
\name{create_case}
\alias{create_case}
\title{Create a base market and a base profile for companies based on case selection criteria.}
\usage{
create_case(
  case = NA,
  number_cost_objects = 3,
  number_cost_pools = 3,
  number_materials = 5,
  number_joint_products = 0,
  start_date = Sys.Date(),
  number_years = 5
)
}
\arguments{
\item{case}{Character vector. Names of the case to select.}

\item{number_cost_objects}{Integer. Number of different produts sold.}

\item{number_cost_pools}{Integer. Number of different cost pools used.}

\item{number_materials}{Integer. Number of different materials used.}

\item{number_joint_products}{Integer. Number of products which are joint.}

\item{start_date}{Date. At which date the first balance sheet is initialized.}

\item{number_years}{Integer. How many years should the time series last.}
}
\value{
list of sublists necessary information for simulation and documentation:
\item{base_market}{parameters about the environment, chart of accounts, resources and market in the case.}
\item{base_company}{parameters about the company, its capacity, technology and underlying costing system, beginning of journal and censuses of various resources.}
}
\description{
Create a base market and a base profile for companies based on case selection criteria.
}
