% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_competition.R
\name{create_competition}
\alias{create_competition}
\title{Create a list of companies based on names and the initial template.}
\usage{
create_competition(company_names, base_company)
}
\arguments{
\item{company_names}{Character vector. names of companies.}

\item{base_company}{List. base_company as returned by the function "create_case".}
}
\value{
list of templates for each company.
}
\description{
Create a list of companies based on names and the initial template.
}
