% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_period.R
\name{create_period}
\alias{create_period}
\title{Create periods for the specified duration.}
\usage{
create_period(date)
}
\arguments{
\item{date}{Date. date of the sale.}
}
\value{
A tibble with different specifications of the period
}
\description{
Create periods for the specified duration.
}
