% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_persons.R
\name{create_persons}
\alias{create_persons}
\title{Prepare figures for small exercises about the high-low method of cost estimation.}
\usage{
create_persons(headcount = 2)
}
\arguments{
\item{headcount}{Numeric. Number of persons to generate.}
}
\value{
A dataframe with gender, ethnicity, first_name, last_name, pronoun, possessive and title.
}
\description{
Prepare figures for small exercises about the high-low method of cost estimation.
}
