% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_batches.R
\name{make_batches}
\alias{make_batches}
\title{Return a new value within plus of minus delta percent of the initial value, adjusting roundings based on the initial magnitude.}
\usage{
make_batches(vol, date, nbatch = 50, shape = 5)
}
\arguments{
\item{vol}{Integer. Volumes to be split in batches.}

\item{date}{Date. Date of the volumes.}

\item{nbatch}{Number of batches to create.}

\item{shape}{Numeric. Positive number; low numbers means skewed to he left, high number to the right.}
}
\value{
A dataframe with lots and their respective volumes.
}
\description{
Return a new value within plus of minus delta percent of the initial value, adjusting roundings based on the initial magnitude.
}
