% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/products_services.R
\docType{data}
\name{products_services}
\alias{products_services}
\title{Consumption of resources per unit of each product}
\format{A data frame with 8 variables:
\itemize{
\item industry
\item family
\item subfamily
\item type
\item item
\item resource
\item unit
\item consumption
}}
\usage{
data("products_services")
}
\description{
Consumption of resources per unit of each product
}
\keyword{datasets}
