% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_allocation.R
\name{record_allocation}
\alias{record_allocation}
\title{Write allocation entries in the ledger.}
\usage{
record_allocation(
  date,
  object,
  allocation_base,
  allocation_rate,
  origin,
  destination
)
}
\arguments{
\item{date}{Date. date of the consumption.}

\item{object}{Character. Name of the materials consumed.}

\item{allocation_base}{Double. Number of units of allocation base.}

\item{allocation_rate}{Double. Amount allocated for each unit of allocation base.}

\item{origin}{Character. From where the consumption is made.}

\item{destination}{Character. Where the consumption goes.}
}
\value{
A tibble of journal entries.
}
\description{
Write allocation entries in the ledger.
}
