% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_consumption.R
\name{record_consumption}
\alias{record_consumption}
\title{Write consumption entries in the ledger.}
\usage{
record_consumption(
  date = Sys.Date(),
  object = "materials",
  quantity = 200,
  price = 3,
  origin = 13300,
  destination = 91000
)
}
\arguments{
\item{date}{Date. date of the consumption.}

\item{object}{Character. Name of the materials consumed.}

\item{quantity}{Integer. Volume consumed.}

\item{price}{Double. Base price of the materials consumed.}

\item{origin}{Character. From where the consumption is made.}

\item{destination}{Character. Where the consumption goes.}
}
\value{
A tibble of journal entries.
}
\description{
Write consumption entries in the ledger.
}
