% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_equity.R
\name{record_equity}
\alias{record_equity}
\title{Write financing entries in the ledger.}
\usage{
record_equity(
  date = Sys.Date(),
  object = "common stock",
  quantity = 100,
  price = 100,
  par = 100,
  origin = 31100,
  destination = 10030
)
}
\arguments{
\item{date}{Date. date of the emission.}

\item{object}{Character. Name of the stock emittedproduced.}

\item{quantity}{Integer. Quantity of stock emitted.}

\item{price}{Double. Market value of the stock emitted.}

\item{par}{Deouble. Book value of the stock emitted.}

\item{origin}{Character. Account where the stock is accounted for.}

\item{destination}{Character. Account where the counterpart is accounted for.}
}
\value{
A tibble of journal entries.
}
\description{
Write financing entries in the ledger.
}
