% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_investment.R
\name{record_investment}
\alias{record_investment}
\title{Write investment entries in the ledger.}
\usage{
record_investment(
  date = Sys.Date(),
  object = "asset",
  price = 120000,
  rate = NA,
  duration = 60,
  origin = 15300,
  destination = 92000
)
}
\arguments{
\item{date}{Date. date of the sale.}

\item{object}{Character. Name of the product or service sold.}

\item{price}{Double. Base price of the product or service.}

\item{rate}{Double. Interest rate used for the operating lease.}

\item{duration}{Integer. Number of periods (months).}

\item{origin}{Character. Account where the asset is "stored".}

\item{destination}{Character. Account where the asset depreciation is accumulated.}
}
\value{
A tibble of journal entries.
}
\description{
Write investment entries in the ledger.
}
