% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_pool_reset.R
\name{record_pool_reset}
\alias{record_pool_reset}
\title{Write pool reset entries in the ledger.}
\usage{
record_pool_reset(date, object, amount, origin, destination)
}
\arguments{
\item{date}{Date. date of the consumption.}

\item{object}{Character. Name of the materials consumed.}

\item{amount}{Double. Amount empryin the cost pool or cost object.}

\item{origin}{Character. From where the consumption is made.}

\item{destination}{Character. Where the consumption goes.}
}
\value{
A tibble of journal entries.
}
\description{
Write pool reset entries in the ledger.
}
