% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_purchase.R
\name{record_purchase}
\alias{record_purchase}
\title{Write purchase entries in the ledger.}
\usage{
record_purchase(
  date = Sys.Date(),
  object = "units",
  quantity = 100,
  price = 10,
  discount = 0.05,
  vat = 0.2,
  dpo = 45,
  risk = 0.1,
  duration = 1,
  origin = 21000,
  destination = 13300
)
}
\arguments{
\item{date}{Date. date of the sale.}

\item{object}{Character. Name of the product or service sold.}

\item{quantity}{Integer. Volume sold.}

\item{price}{Double. Base price of the product or service.}

\item{discount}{Double. Percentage of price reduction earned if payment within terms.}

\item{vat}{Double. Percentage of price to compute value added taxes.}

\item{dpo}{Integer. Day-Payable-Outstading or customer credit.}

\item{risk}{Double. Probability of not paying on time.}

\item{duration}{Integer. For how many periods does the right hold.}

\item{origin}{Character. From where the purchase is made.}

\item{destination}{Character. Where the purchase goes.}
}
\value{
A tibble of journal entries.
}
\description{
Write purchase entries in the ledger.
}
