% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_purchases.R
\name{record_purchases}
\alias{record_purchases}
\title{Write entries in the ledger related to purchases.}
\usage{
record_purchases(
  date = Sys.Date(),
  object = "units",
  quantity = 100,
  price = 10,
  discount = 0.05,
  vat = 0.2,
  dpo = 45,
  risk = 0.1,
  lifetime = 1,
  nature = 21000,
  destination = 13300
)
}
\arguments{
\item{date}{Date. date of the sale.}

\item{object}{Character. Name of the product or service sold.}

\item{quantity}{Integer. Volume sold.}

\item{price}{Double. Base price of the product or service.}

\item{discount}{Double. Percentage of price reduction earned if payment within terms.}

\item{vat}{Double. Percentage of price to compute value added taxes.}

\item{dpo}{Integer. Day-Payable-Outstading or customer credit.}

\item{risk}{Double. Probability of not paying on time.}

\item{lifetime}{Integer. For how many periods does the right hold.}

\item{nature}{Character. Account where the asset is "stored".}

\item{destination}{Character. Account where the asset depreciation is accumulated.}
}
\value{
A tibble of journal entries.
}
\description{
Write entries in the ledger related to purchases.
}
