% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_salary.R
\name{record_salary}
\alias{record_salary}
\title{Write salary entries in the ledger.}
\usage{
record_salary(
  date = Sys.Date(),
  object = "salespeople",
  quantity = 1,
  price = 2000,
  rate = 0.5,
  origin = 22100,
  destination = 61200
)
}
\arguments{
\item{date}{Date. date of the recognition and payment.}

\item{object}{Character. Name of the kind of employees.}

\item{quantity}{Integer. Time paid in Full Time Equivalent.}

\item{price}{Double. Monthly salary of the workfroce.}

\item{rate}{Double. Percentage of base salary to compute labor taxes.}

\item{origin}{Character. From where the salary is taken.}

\item{destination}{Character. Where the salary goes.}
}
\value{
A tibble of journal entries.
}
\description{
Write salary entries in the ledger.
}
