% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_sale.R
\name{record_sale}
\alias{record_sale}
\title{Write sales entries in the journal.}
\usage{
record_sale(
  date = Sys.Date(),
  object = "units",
  quantity = 100,
  price = 10,
  discount = 0.05,
  vat = 0.2,
  dso = 30,
  unit_cost = 5,
  sales_commission = 0.1,
  labor_tax = 0.5,
  dco = 15,
  account_increment = 0,
  situation = NULL,
  risk = 0.1
)
}
\arguments{
\item{date}{Date. date of the sale.}

\item{object}{Character. Name of the product or service sold.}

\item{quantity}{Integer. Volume sold.}

\item{price}{Double. Base price of the product or service.}

\item{discount}{Double. Percentage of price reduction earned if payment within terms.}

\item{vat}{Double. Percentage of price to compute value added taxes.}

\item{dso}{Integer. Day-Sales-Outstading or customer credit.}

\item{unit_cost}{Double. Cost of the product or services sold.}

\item{sales_commission}{Double. Percentage of price paid to sales people as commission (end of month).}

\item{labor_tax}{Double. Percentage of labor taxes applied to the commissions.}

\item{dco}{Integer. Delay before paying commissions.}

\item{account_increment}{Integer. Last two digits of an account number indicating the destination.}

\item{situation}{Character. NULL or specify one of the following situation: "default", "sales return", "pay very late" or "pay late"}

\item{risk}{Double. Number between 0 and 1 indicating the risk on the sale, affecting the probability of different situations.}
}
\value{
A tibble of journal entries.
}
\description{
Write sales entries in the journal.
}
