% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_tax_payment.R
\name{record_tax_payment}
\alias{record_tax_payment}
\title{Write tax payment entries in the ledger.}
\usage{
record_tax_payment(date, object, amount, origin)
}
\arguments{
\item{date}{Date. date of the consumption.}

\item{object}{Character. Name of the materials consumed.}

\item{amount}{Double. Amount due to the State.}

\item{origin}{Character. From where the consumption is made.}
}
\value{
A tibble of journal entries.
}
\description{
Write tax payment entries in the ledger.
}
