% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\docType{data}
\name{resources}
\alias{resources}
\title{Unit price for each resource}
\format{A data frame with 4 variables
\itemize{
\item resource
\item nature
\item unit
\item cost
}}
\usage{
data("resources")
}
\description{
Unit price for each resource
}
\keyword{datasets}
