% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_cases.R
\name{select_cases}
\alias{select_cases}
\title{Functions generating cases for exercises}
\usage{
select_cases(
  product = NULL,
  family = NULL,
  nbr_products = 1,
  mono_material = FALSE,
  joint = FALSE,
  nbr_companies = 1
)
}
\arguments{
\item{product}{Character vector. Names of the product to select.}

\item{family}{Character. Name of the family to select.}

\item{nbr_products}{Integer. Number of different products.}

\item{mono_material}{Logical. Whether the selected products should be based on a unique material.}

\item{joint}{Logical. Whether the selected products should be joint products.}

\item{nbr_companies}{Integer. Number of different companies.}
}
\value{
list of sublists necessary information for simulation and documentation:
\item{description}{Stable information about "families", "products", "activities", "resources", "contracts" and "consumptions.}
\item{environment}{Parameters controlled by the game master: "seasons", "parameters" and "contract_premia".}
\item{competition}{Parameters controlled or influenced by the players at the following levels: "company", "product", "resource" and "product_resource"}
}
\description{
Functions generating cases for exercises
}
