% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul_attractiveness.R
\name{simul_attractiveness}
\alias{simul_attractiveness}
\title{Compute attractiveness scores of references based on prime costs, advertising and price.}
\usage{
simul_attractiveness(
  profiles,
  sensitivity_prime_cost = 2,
  sensitivity_advertising = 2,
  sensitivity_price = 3,
  sensitivity_dso = 1
)
}
\arguments{
\item{profiles}{Tibble. One observation for each reference (company-product) on the market, and the following properties: "reference","prime_cost","advertising", "price" and "dso".}

\item{sensitivity_prime_cost}{Numeric. Relative weight customers put on prime cost.}

\item{sensitivity_advertising}{Numeric. Relative weight customers put on prime advertising.}

\item{sensitivity_price}{Numeric. Relative weight customers put on prime price.}

\item{sensitivity_dso}{Numeric. Relative weight customers put on DSO (credit terms).}
}
\value{
A tibble with the relative attractiveness of each reference
}
\description{
Compute attractiveness scores of references based on prime costs, advertising and price.
}
\seealso{
simul_market
}
