% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul_batches.R
\name{simul_batches}
\alias{simul_batches}
\title{Return a new value within plus of minus delta percent of the initial value, adjusting roundings based on the initial magnitude.}
\usage{
simul_batches(vol, nbatch = 50, shape = 5)
}
\arguments{
\item{vol}{Integer. Volumes to be split in batches.}

\item{nbatch}{Number of batches to create.}

\item{shape}{Numeric. Positive number for the gamma distribution; low numbers means skewed to he left, high number to the right.}
}
\value{
A dataframe with lots and their respective volumes.
}
\description{
Return a new value within plus of minus delta percent of the initial value, adjusting roundings based on the initial magnitude.
}
