% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul_demand.R
\name{simul_demand}
\alias{simul_demand}
\title{Simulate a base market over several periods based on a set of assumptions.}
\usage{
simul_demand(market_base = 1000, profiles = NULL, demand_function = "logistic")
}
\arguments{
\item{market_base}{Integer. Base volume of the demand.}

\item{profiles}{Tibble. One row per reference (company-product), its name and its attractiveness (output of the function simul_attractiveness).}

\item{demand_function}{Character. Whether the demand is a "linear", "constant" elasiticy, or "logistic" function of price.}
}
\value{
A volume of the demand to the firm.
}
\description{
Simulate a base market over several periods based on a set of assumptions.
}
\seealso{
simul_market
}
