% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul_market.R
\name{simul_market}
\alias{simul_market}
\title{Simulate a base market over several periods based on a set of assumptions.}
\usage{
simul_market(start = Sys.Date(), years = 5, base_volume = 1000,
  trend = 0.05, randomness = 0.1, coeffday = NULL,
  coeffmonth = NULL)
}
\arguments{
\item{start}{Character or date. When the the time series should start.}

\item{years}{Integer. How many years should the time series last.}

\item{base_volume}{Integer. Set magnitude for average daily firm demand.}

\item{trend}{Double. Linear trader across all periods.}

\item{randomness}{Double. Percentage of random variation for the demand.}

\item{coeffday}{Tibble. 7 observations for 2 variables: the week day (short) and the distribution (higher numbers indicate on which days demand should be concentrated)}

\item{coeffmonth}{Tibble. 12 observations for 2 variables: the month (number) and the distribution (higher numbers indicate on which months demand should be concentrated)}
}
\value{
A tibble with the market size for eachh period.
}
\description{
Simulate a base market over several periods based on a set of assumptions.
}
