% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_period.R
\name{simulate_period}
\alias{simulate_period}
\title{Simulate the specified period of activity.}
\usage{
simulate_period(case, period_nbr)
}
\arguments{
\item{case}{List. output of either create_case of simulate_period.}

\item{period_nbr}{Integer. Number (i.d. order) of the period to simulate.}
}
\value{
Append updated journal and census to the competitors' parameters.

list of sublists necessary information for simulation and documentation:
\item{base_market}{parameters about the environment, chart of accounts and market in the case.}
\item{base_company}{parameters about the company, its capacity, technology, beginning of journal and censuses of various resources.}
}
\description{
Simulate the specified period of activity.
}
