% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_activity.R
\name{update_activity}
\alias{update_activity}
\title{Update information about companies' competitive profiles in competition.}
\usage{
update_activity(competition, simperiod, simworkdays)
}
\arguments{
\item{competition}{List. competitors as returned by the function make_competition.}

\item{simperiod}{Character. ID of the period for which the profile holds.}

\item{simworkdays}{Integer. Number of working days in the period.}
}
\value{
Append updated activity to the competitors' parameters.
}
\description{
Update information about companies' competitive profiles in competition.
}
