% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_expenses.R
\name{update_expenses}
\alias{update_expenses}
\title{Update information about companies' journal and census based on operations}
\usage{
update_expenses(competition, simperiod, simworkdays, base_market)
}
\arguments{
\item{competition}{List. competitors as returned by the function make_competition.}

\item{simperiod}{Character. ID of the period for which the profile holds.}

\item{simworkdays}{Integer. Number of working days in the period.}

\item{base_market}{List. market based returned by the function create_case.}
}
\value{
Append updated journal and census to the competitors' parameters.
}
\description{
Update information about companies' journal and census based on operations
}
