% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_profiles.R
\name{update_profiles}
\alias{update_profiles}
\title{Update information about companies' competitive profiles in competition.}
\usage{
update_profiles(simperiod, simdemand, competition, base_market)
}
\arguments{
\item{simperiod}{Character. ID of the period for which the profile holds.}

\item{simdemand}{Integer. Demand for the period.}

\item{competition}{List. competitors as returned by the function make_competition.}

\item{base_market}{List. market based returned by the function create_case.}
}
\value{
Append updated profiles to the competitors' parameters.
}
\description{
Update information about companies' competitive profiles in competition.
}
