% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiggle.R
\name{wiggle}
\alias{wiggle}
\title{Return a new value within plus of minus delta percent of the initial value, adjusting roundings based on the initial magnitude.}
\usage{
wiggle(x, delta = 0.2, dir = 0)
}
\arguments{
\item{x}{Numeric. Value around which to wiggle}

\item{delta}{Numeric. percentage of variation around the initial value.}

\item{dir}{Numeric. -1 for negative, 1 for positive, 0 for either.}
}
\value{
A new value
}
\description{
Return a new value within plus of minus delta percent of the initial value, adjusting roundings based on the initial magnitude.
}
