% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_companies.R
\docType{data}
\name{case_companies}
\alias{case_companies}
\title{List of general parameters for a simulation.}
\format{List of general parameters for a simulation.
\itemize{
\item case: id of the case.
\item company_type: indication about the profile of the company
\item number_cost_objects: number of different produts or services sold.
\item number_cost_pools: number of different cost pools used.
\item number_materials: number of different kinds of raw materials used.
\item number_joint_products: number of products which are joint.
\item desired_rate_of_return: minimum rate of return on assets desired.
\item interest_rate: initial interest rate.
\item overdraft_rate: interest rate applied on advances from the bank.
\item cost_of_capital: rate of return required by shareholders.
\item debt_share: proportion of the capital employed financed by debt at the beginning.
\item debt_years: number of years over which the debt was contracted.
\item cash_buffer: percentage of total initial investment added as cash to finance aoperations.
\item risk_customer: risk of an issue with collection.
\item risk_purchaser: risk of an issue (losing a discount) with payments
}}
\usage{
data("case_companies")
}
\description{
List of general parameters for a simulation.
}
\keyword{datasets}
