% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_seasons.R
\docType{data}
\name{case_seasons}
\alias{case_seasons}
\title{Relative distribution of the demand over weeks and weekdays for one year.}
\format{Relative distribution of the demand over weeks and weekdays for one year.
\itemize{
\item case: id of the case.
\item week: week number
\item monday: demand size on Mondays.
\item tuesday: demand size on Tuesdays.
\item wednesday: demand size on Wednesdays.
\item thursday: demand size on Thursdays.
\item friday: demand size on Fridays.
\item saturday: demand size on Saturdays.
\item sunday: demand size on Sundays.
}}
\usage{
data("case_seasons")
}
\description{
Relative distribution of the demand over weeks and weekdays for one year.
}
\keyword{datasets}
