% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_case.R
\name{create_case}
\alias{create_case}
\title{Create a base market and a base profile for companies based on case selection criteria.}
\usage{
create_case(
  case = NULL,
  start_date = Sys.Date(),
  number_months = 12,
  number_companies = 1
)
}
\arguments{
\item{case}{List. Tibbles about case information, environments, seasons, accounts, capacity, technology and costing.}

\item{start_date}{Date. At which date the first balance sheet is initialized.}

\item{number_months}{Integer. How many months should the time series last.}

\item{number_companies}{Integer. How many companies should be created.}
}
\value{
list of sublists necessary information for simulation and documentation:
\item{base_market}{parameters about the environment, chart of accounts and market in the case.}
\item{base_company}{parameters about the company, its capacity, technology, beginning of journal and censuses of various resources.}
}
\description{
Create a base market and a base profile for companies based on case selection criteria.
}
