% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_case.R
\name{create_case}
\alias{create_case}
\title{Functions generating cases for exercises}
\usage{
create_case(
  case = NA,
  number_cost_objects = 3,
  number_cost_pools = 3,
  number_materials = 5,
  number_production_steps = 1,
  number_joint_products = 0,
  start_date = Sys.Date()
)
}
\arguments{
\item{case}{Character vector. Names of the case to select.}

\item{number_cost_objects}{Integer. Number of different produts sold.}

\item{number_cost_pools}{Integer. Number of different cost pools used.}

\item{number_materials}{Integer. Number of different materials used.}

\item{number_production_steps}{Integer. Number of sequential production steps.}

\item{number_joint_products}{Integer. Number of products which are joint.}

\item{start_date}{Date. At which date the first balance sheet is initialized.}
}
\value{
list of sublists necessary information for simulation and documentation:
\item{base_market}{parameters about the environment, seasonality, chart of accounts and periods in the case.}
\item{company}{parameters about the company, its capacity, activity, costing system, inventories and beginning of journal.}
}
\description{
Functions generating cases for exercises
}
