% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul_demand.R
\name{simul_demand}
\alias{simul_demand}
\title{Simulate a base market over several periods based on a set of assumptions.}
\usage{
simul_demand(market_base = 1000, prices = c(25, 50, 75, 100),
  qualities = c(20, 55, 65, 105), sensitivity = 0.5,
  demand_function = "logistic")
}
\arguments{
\item{market_base}{Integer. Base volume of the demand.}

\item{prices}{Numeric vector. Prices applied for each product.}

\item{qualities}{Numeric vector. Relative qualities of each product.}

\item{sensitivity}{Numeric. Number between 0 (less sensitive) and 1 (more sensitive) indicating how sentitive to quality price demand is.}

\item{demand_function}{Character. Whether the demand is a "linear", "constant" elasiticy, or "logistic" function of price.}
}
\value{
A volume of the demand to the firm.
}
\description{
Simulate a base market over several periods based on a set of assumptions.
}
\seealso{
simul_market
}
